wordnet_autotranslate/integrations/dspy_integration.py
"""DSPy integration with NWO format support."""

import logging
from typing import Dict, List, Optional, Any
from ..config import Config, NWOProvider

logger = logging.getLogger(__name__)


class DSPyIntegration:
    """Integration with DSPy framework supporting NWO format."""
    
    def __init__(self, config: Config):
        """Initialize DSPy integration.
        
        Args:
            config: Configuration object
        """
        self.config = config
        self.dspy_config = config.get_dspy_config()
        self._setup_dspy()
    
    def _setup_dspy(self) -> None:
        """Set up DSPy framework."""
        try:
            import dspy
            self.dspy = dspy
            
            # Configure DSPy based on NWO providers
            nwo_providers = {
                name: provider 
                for name, provider in self.config.providers.items()
                if provider.namespace == 'dspy'
            }
            
            logger.info(f"Initialized DSPy integration with {len(nwo_providers)} NWO providers")
            
        except ImportError:
            logger.warning("DSPy not available - install with: pip install dspy-ai")
            self.dspy = None
    
    def translate(
        self, 
        text: str, 
        source_lang: str, 
        target_lang: str, 
        provider: Optional[NWOProvider] = None
    ) -> str:
        """Translate text using DSPy with WordNet enhancement.
        
        Args:
            text: Text to translate
            source_lang: Source language code
            target_lang: Target language code
            provider: NWO provider specification
            
        Returns:
            Translated text
        """
        if not self.dspy:
            raise RuntimeError("DSPy not available")
        
        # Get WordNet synsets for better translation
        synsets = self.get_synsets(text.split(), source_lang)
        
        # Enhanced translation using WordNet context
        enhanced_text = self._enhance_with_wordnet(text, synsets)
        
        # Use DSPy for translation (mock implementation)
        translated = self._dspy_translate(enhanced_text, source_lang, target_lang, provider)
        
        return translated
    
    def _enhance_with_wordnet(self, text: str, synsets: List[Dict[str, Any]]) -> str:
        """Enhance text with WordNet semantic information.
        
        Args:
            text: Original text
            synsets: WordNet synsets for context
            
        Returns:
            Enhanced text with semantic annotations
        """
        # For now, return original text - in a real implementation,
        # this would add semantic context from synsets
        logger.debug(f"Enhancing text with {len(synsets)} synsets")
        return text
    
    def _dspy_translate(
        self, 
        text: str, 
        source_lang: str, 
        target_lang: str, 
        provider: Optional[NWOProvider]
    ) -> str:
        """Perform translation using DSPy framework.
        
        Args:
            text: Text to translate
            source_lang: Source language
            target_lang: Target language
            provider: NWO provider specification
            
        Returns:
            Translated text
        """
        # Mock DSPy translation - in a real implementation,
        # this would use DSPy's language models
        provider_info = f" via {provider.to_string()}" if provider else ""
        logger.info(f"DSPy translating '{text}' from {source_lang} to {target_lang}{provider_info}")
        
        # Simple mock translation for demonstration
        if target_lang == 'es' and source_lang == 'en':
            # Mock Spanish translation
            translations = {
                'hello': 'hola',
                'world': 'mundo',
                'cat': 'gato',
                'dog': 'perro',
                'house': 'casa',
                'book': 'libro',
            }
            
            words = text.lower().split()
            translated_words = []
            
            for word in words:
                # Remove punctuation for lookup
                clean_word = ''.join(c for c in word if c.isalnum())
                if clean_word in translations:
                    # Preserve original punctuation
                    translated = word.replace(clean_word, translations[clean_word])
                    translated_words.append(translated)
                else:
                    translated_words.append(word)
            
            return ' '.join(translated_words)
        
        # Default: return original text with language annotation
        return f"[{target_lang}] {text}"
    
    def get_synsets(self, words: List[str], lang: str = 'en') -> List[Dict[str, Any]]:
        """Get WordNet synsets for words using DSPy approach.
        
        Args:
            words: List of words to look up
            lang: Language code
            
        Returns:
            List of synset information
        """
        synsets = []
        
        for word in words:
            # Mock synset data - in a real implementation,
            # this would query WordNet through DSPy
            synset_data = {
                'word': word,
                'lang': lang,
                'synsets': [
                    {
                        'name': f"{word}.n.01",
                        'definition': f"Mock definition for {word}",
                        'examples': [f"Example usage of {word}"],
                        'hypernyms': [],
                        'hyponyms': [],
                        'synonyms': [],
                    }
                ],
                'provider': '@dspy/wordnet'
            }
            synsets.append(synset_data)
        
        logger.debug(f"Retrieved {len(synsets)} synsets for {len(words)} words via DSPy")
        return synsets
    
    def get_supported_languages(self) -> List[str]:
        """Get list of languages supported by DSPy integration.
        
        Returns:
            List of language codes
        """
        # Mock supported languages - in a real implementation,
        # this would query DSPy's capabilities
        return ['en', 'es', 'fr', 'de', 'it', 'pt', 'ru', 'zh', 'ja', 'ko']
    
    def get_model_info(self, provider: Optional[NWOProvider] = None) -> Dict[str, Any]:
        """Get information about the DSPy model being used.
        
        Args:
            provider: NWO provider specification
            
        Returns:
            Model information
        """
        provider_str = provider.to_string() if provider else "default"
        
        return {
            'provider': provider_str,
            'framework': 'DSPy',
            'model_type': 'language_model',
            'wordnet_integration': True,
            'supports_nwo': True,
            'version': getattr(self.dspy, '__version__', 'unknown') if self.dspy else 'unavailable'
        }