requirements.txt
nltk>=3.8
dspy-ai>=2.0.0
numpy>=1.21.0
PyYAML>=6.0
requests>=2.25.0